backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
networkx==3.2.1
numpy>=1.24.0
scikit-learn>=1.4.0
pandas>=2.0.0
matplotlib>=3.7.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0